mooved schema to separate file
